Pause for Bus interface implementation
This approach is one I discovered through some poking around online
forums. The one constraint to this interface is no one can implement
their own Bus, as the entire union is already set in stone.

However, I am not building a library, so this is good for me. Now the
CPU Tests run on a simple Bus and can begin work on the Bus for the
"production" emulator.